24-Hr-Hackathon/ai-service/requirements.txt
flask==2.3.2
flask-cors==4.0.0
opencv-python==4.8.0.74
numpy==1.24.3
Pillow==9.5.0